DDL replication for Table DDL commands
OVERVIEW
--------
1) To support DDL replication, we use the event trigger and DDL deparsing
facilities. During CREATE PUBLICATION we register a 'command end' trigger that
deparses the DDL (if the DDL is annotated as ddlreplok for DDL replication in
cmdtaglist.h) as a JSON blob, and WAL logs it. The event trigger is automatically
removed at the time of DROP PUBLICATION. The WALSender decodes the WAL and sends
it downstream similar to other DML commands. The subscriber then converts JSON
back to the DDL command string and executes it. In the subscriber, we also add
the newly added rel to pg_subscription_rel so that the DML changes on the new
table can be replicated without having to manually run
"ALTER SUBSCRIPTION ... REFRESH PUBLICATION".

2) Add subscription tap test for DDL replication for TABLE related commands.

For non-rewrite ALTER object command and CREATE object command:
---------------------------------------------------------------
We deparse the command at ddl_command_end event trigger and WAL log the
deparsed JSON string. The WALSender decodes the WAL and sends it to
subscriber if the created/altered table is published.

For DROP object:
----------------
The 'command start' event handler logs a ddl message with the relids of
the tables that are dropped which the output plugin (pgoutput) stores in
its internal data structure after verifying that it is for a table that is
part of the publication. Later the 'command end' event handler sends the
actual drop message. On receiving the 'command end', pgoutput sends the
DROP command only if it is for one of the relids marked for deletion.

The reason we have to do this is because, once the logical decoder
receives the 'command end' message,  the relid of the table is no longer
valid as it has been deleted as part of invalidations received for the
drop table command. It is no longer possible to verify if the table is
part of the publication list or not. To make this possible, two more elements
are added to the ddl xlog and ddl message, (relid and cmdtype).

We could have also handled all this on the subscriber side, but
that would mean sending spurious ddl messages for tables that are not part
of the publication.

For table_rewrite ALTER TABLE command:
--------------------------------------
Executing a non-immutable expression during the table rewrite phase is not
allowed, as it may result in different data between publisher and subscriber.
While some may suggest converting the rewrite inserts to updates and replicate
them afte the ddl command to maintain data consistency. But it doesn't work if
the replica identity column is altered in the command. This is because the
rewrite inserts do not contain the old values and therefore cannot be converted
to update.

LIMIT:
Commands contain volatile functions are not allowed. Because
it's possible the functions contain DDL/DML in which case these operations
will be executed twice and cause duplicate data. In addition, we don't know
whether the tables being accessed by these DDL/DML are published or not. So
blindly allowing such functions can allow unintended clauses like the tables
accessed in those functions may not even exist on the subscriber.