test deparser module regress twice
When running the regression test(tests in parallel_schedule), we replace the
executing ddl statement with the its deparsed version and execute the
deparsed statement, so that we can run all the regression with the deparsed
statement and can expect the output to be the same as the existing
expected/*.out. As developers typically add new regression tests to test new
syntax, so we expect this test can automatically identify any new syntax
changes.

To achieve this, the strategy is to run the regression test twice. The first
run will create event triggers to deparse ddl statements, it is intended to
collect all the deparsed statements and can catch ERRORs/WARNINGs caused by
the deparser. We can dump these statements from the database and then reload
them in the second regression run. This allows us to utilize the deparsed statements
to replace the local statements in the second regression run. This approach
does not need to handle any remote messages and client variable stuff during
execution, although it could take more time to finsh the test.